reorganized shaders
