+++ src/components/Canopy.tsx
@@ -6,8 +6,8 @@ import {
 } from "three";
 import { useFrame, useThree } from "@react-three/fiber";
 import { useEffect, useMemo, useRef } from "react";
-import canopyVert from "@/src/patterns/shaders/canopy.vert";
-import fromTextureWithIntensity from "@/src/patterns/shaders/fromTextureWithIntensity.frag";
+import canopyVert from "@/src/shaders/canopy.vert";
+import fromTextureWithIntensity from "@/src/shaders/fromTextureWithIntensity.frag";
 import canopyGeometry from "@/src/data/canopyGeometry.json";
 import {
   BloomEffect,
