+++ src/shaders/fromTextureWithIntensity.frag
